% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem Simulation for Uniform Distribution}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{Sample size for each iteration}

\item{iter}{Number of iterations/samples}

\item{a}{Lower bound of uniform distribution (default = 0)}

\item{b}{Upper bound of uniform distribution (default = 5)}
}
\value{
A histogram with a normal distribution overlay; returns the vector of simulated sums
}
\description{
Simulates the CLT using uniform(a, b) distributions. Repeatedly samples and sums random values.
}
