% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Bootstrap Confidence Interval Estimator}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A likelihood function that accepts data and a parameter.}

\item{x}{A vector of observed data.}

\item{param}{A vector of parameter values over which the likelihood is evaluated.}

\item{...}{Additional arguments passed to the plot function.}
}
\value{
List with i, parami, yi, and slope
}
\description{
Bootstrap Confidence Interval Estimator
}
