% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Mysample Function}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{A quantitative vector that represents}

\item{iter}{A quantitative vector}

\item{time}{A quantitative vector}
}
\value{
A quantitative vector
}
\description{
Mysample Function
}
\examples{
mysample(n=1000, iter=30,time=1)
}
