% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Mysample Function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{A quantitative vector which represents how many seats are available on the flight}

\item{gamma}{A quantitative vector that represents the confidence interval or the probability of over booking a flight}

\item{p}{A quantitative vector that represents the probability of a individual passenger showing up for the flight}
}
\value{
A list containing:
\item{nd}{The optimal number of tickets based on the binomial distribution (discrete).}
\item{nc}{The optimal number of tickets based on the approximate normal distribution (continuous).}
\item{gamma}{The probability of over booking the flight, which was imputed.}
\item{p}{The probability of an individual showing up for flighty, which was imputed.}
\item{N}{The number of seats on flighty, which was imputed.}
}
\description{
Mysample Function
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
